西元2023年08月08日 (週二) 18時32分02秒 CST

+++ app/ub/items/CaptionDownloader/CaptionDownloader.js
@@ -6,30 +6,49 @@ const fs = require('fs')
 
 const TimeMarkAnalysis = require('./TimeMarkAnalysis.js')
 
+let isRunning = false
+let sleep = function (ms = 500) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
+
 // https://youtu.be/85AqJsmxDZs
 module.exports = async function (utID = 'https://youtu.be/JxbFotMgqik', timeMarkList = []) {
   let expire = 365 * 24 * 60 * 60 * 1000
-  expire = 10  // for test
+  // expire = 10  // for test
 
-  // return await NodeCacheSqlite.get('CaptionDownloader', utID, async () => {
-    if (utID.startsWith('http')) {
-      utID = UBVideoIdParser(utID)
+  return await NodeCacheSqlite.get('CaptionDownloader', utID, async () => {
+    try {
+      if (utID.startsWith('http')) {
+        utID = UBVideoIdParser(utID)
+      }
+
+      while (isRunning === true) {
+        await sleep(1000)
+      }
+
+      isRunning = true
+      await ShellSpawn([`python3`, `/app/python/caption.py`, utID])
+      isRunning = false
+      let srtPath = `/app/tmp/srt-${utID}.txt`
+      if (fs.existsSync(srtPath) === false) {
+        return false
+      }
+      let srt = fs.readFileSync(srtPath, 'utf-8')
+
+      if (Array.isArray(timeMarkList) === false) {
+        timeMarkList = TimeMarkAnalysis(timeMarkList)
+      }
+
+      captionParagraph = CaptionFormat(srt, timeMarkList)
+      // console.log(captionParagraph)
+
+      fs.unlinkSync(srtPath)
+
+      return captionParagraph
     }
-
-    await ShellSpawn([`python3`, `/app/python/caption.py`, utID])
-    let srtPath = `/app/tmp/srt-${utID}.txt`
-    let srt = fs.readFileSync(srtPath, 'utf-8')
-
-    if (Array.isArray(timeMarkList) === false) {
-      timeMarkList = TimeMarkAnalysis(timeMarkList)
+    catch (e) {
+      return false
     }
-
-    let captionParagraph = CaptionFormat(srt, timeMarkList)
-    // console.log(captionParagraph)
-
-    fs.unlinkSync(srtPath)
-
-    return captionParagraph
-  // }, expire)
+  }, expire)
     
 }
